pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
my-script = "prvniProject.main:main"

[tool.poetry]
name = "prvniProject"
version = "0.1.0"
description = ""
authors = ["Ken <ken@somewhere.io>"]
readme = "README.md"
packages = [{include = "prvniProject"}]

[tool.poetry.dependencies]
python = "^3.7"
mysql-connector-python = "8.0.33"
SQLAlchemy = "2.0"
flask = "2.3.2"

[tool.poetry.scripts]
prvniProject = 'prvniProject.main:main'